+++ app/src/main/java/com/example/stanley_kubrick_archive/MainActivity.kt
@@ -52,6 +52,7 @@ import androidx.compose.ui.unit.dp
 import com.example.stanley_kubrick_archive.component.ActorCardItem
 import com.example.stanley_kubrick_archive.component.BottomGradientLayer
 import com.example.stanley_kubrick_archive.component.MovieItem
+import com.example.stanley_kubrick_archive.component.MoviesCardsList
 import com.example.stanley_kubrick_archive.component.ToolbarWithTextAndIcon
 import com.example.stanley_kubrick_archive.data.MovieCard
 import com.example.stanley_kubrick_archive.data.movieList
@@ -88,165 +89,4 @@ class MainActivity : ComponentActivity() {
 @Composable
 private fun HomeScreen(items: List<MovieCard>, selectedCard: MutableState<Int?>) {
     MoviesCardsList(items = items, selectedCard)
-}
-
-
-@OptIn(ExperimentalComposeUiApi::class, ExperimentalFoundationApi::class)
-@Composable
-fun MoviesCardsList(items: List<MovieCard>, selectedCard: MutableState<Int?>) {
-    val listState = rememberLazyListState()
-    val pagerState = rememberPagerState()
-    val animationInProgress = remember { mutableStateOf<Boolean>(false) }
-    val userScrollEnabled = remember { mutableStateOf<Boolean>(true) }
-    Box {
-        LazyColumn(
-            modifier = Modifier.padding(horizontal = 30.dp),
-            state = listState,
-            verticalArrangement = Arrangement.spacedBy((-124).dp),
-            userScrollEnabled = userScrollEnabled.value,
-        ) {
-            itemsIndexed(items) { index, item ->
-                MovieItem(
-                    item,
-                    index,
-                    listState,
-                    selectedCard,
-                    animationInProgress,
-                    userScrollEnabled,
-                    pagerState
-                )
-            }
-        }
-        //DetailsMovie(selectedCard, listState)
-
-        AnimatedVisibility(visible = selectedCard.value != null) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(top = 300.dp)
-            ) {
-                HorizontalPager(
-                    pageCount = 3,
-                    state = pagerState,
-                    modifier = Modifier
-                        .weight(1f)
-                        .fillMaxWidth()
-                ) {
-
-                        page ->
-                    Box(
-                        modifier = Modifier
-                            .fillMaxSize()
-                            .background(Color(0xFF, 0x80 * (page + 1), 0x80)),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Text(text = "Page $page", color = Color.White)
-                    }
-                }
-                PagerIndicator(
-                    pagerState = pagerState,
-                    pageCount = 3,
-                    modifier = Modifier.align(Alignment.CenterHorizontally)
-                )
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun PagerIndicator(pagerState: PagerState, pageCount: Int, modifier: Modifier = Modifier) {
-    Row(modifier = modifier) {
-        repeat(pageCount) { index ->
-            Box(
-                modifier = Modifier
-                    .size(if (pagerState.currentPage == index) 10.dp else 8.dp)
-                    .background(if (pagerState.currentPage == index) Color.Red else Color.Gray)
-                    .padding(horizontal = 4.dp)
-            )
-        }
-    }
-}
-
-@Composable
-fun ExpandingCard() {
-    var isExpanded by remember { mutableStateOf(false) }
-    val transition = updateTransition(targetState = isExpanded, label = "cardTransition", )
-    val cardSize = animateDpAsState(targetValue = if (isExpanded) 120.dp else 240.dp, label = "cardSize", animationSpec = tween<Dp>(
-        durationMillis = 3000,
-        easing = LinearEasing
-    ) )
-    val cardPadding = animateDpAsState(targetValue = if (isExpanded) 8.dp else 0.dp, label = "cardPadding", animationSpec = tween<Dp>(
-        durationMillis = 3000,
-        easing = LinearEasing
-    ))
-
-    Box(contentAlignment = Alignment.Center, modifier = Modifier.fillMaxSize()) {
-        Column(modifier = Modifier
-            .clickable { isExpanded = !isExpanded }
-            .padding(all = cardPadding.value)
-
-            .animateContentSize(),
-            horizontalAlignment = Alignment.CenterHorizontally) {
-
-            if (isExpanded) {
-                Row {
-                    ActorCardItem(imageRes = R.drawable.clockwork_orange)
-                    ActorCardItem(imageRes = R.drawable.bary)
-                }
-                Row {
-                    ActorCardItem(imageRes = R.drawable.shining)
-                    ActorCardItem(imageRes = R.drawable.i)
-                }
-            } else {
-                Card(modifier = Modifier.size(240.dp)) {
-                    Image(
-                        painter = painterResource(R.drawable.eyes_wide_shut),
-                        contentDescription = null,
-                        modifier = Modifier.fillMaxSize(),
-                        contentScale = ContentScale.Crop
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun SmallCard(imageRes: Int) {
-    Card(
-        modifier = Modifier
-            .padding(4.dp)
-            .size(100.dp),
-
-    ) {
-        AnimatedVisibility(
-            visible = true,
-            enter = expandIn(),
-            exit = shrinkOut()
-        ) {
-            Image(
-                painter = painterResource(imageRes),
-                contentDescription = null,
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.Crop
-            )
-        }
-    }
-}
-
-suspend fun PointerInputScope.detectVerticalSwipeGestures(
-    onSwipeUp: (dragAmount: Float) -> Unit,
-    onSwipeDown: (dragAmount: Float) -> Unit,
-) {
-    detectDragGestures { change, dragAmount ->
-        change.consumeAllChanges()
-        if (abs(dragAmount.y) > abs(dragAmount.x)) { // Detect vertical movement
-            if (dragAmount.y < 0) {
-                onSwipeUp(dragAmount.y)
-            } else {
-                onSwipeDown(dragAmount.y)
-            }
-        }
-    }
 }
