+++ app/src/main/java/com/example/stanley_kubrick_archive/component/MoviesListComponent.kt
@@ -1,0 +1,88 @@
+package com.example.stanley_kubrick_archive.component
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.pager.HorizontalPager
+import androidx.compose.foundation.pager.rememberPagerState
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import com.example.stanley_kubrick_archive.PagerIndicator
+import com.example.stanley_kubrick_archive.data.MovieCard
+
+@OptIn(ExperimentalComposeUiApi::class, ExperimentalFoundationApi::class)
+@Composable
+fun MoviesCardsList(items: List<MovieCard>, selectedCard: MutableState<Int?>) {
+    val listState = rememberLazyListState()
+    val pagerState = rememberPagerState()
+    val animationInProgress = remember { mutableStateOf<Boolean>(false) }
+    val userScrollEnabled = remember { mutableStateOf<Boolean>(true) }
+    Box {
+        LazyColumn(
+            modifier = Modifier.padding(horizontal = 30.dp),
+            state = listState,
+            verticalArrangement = Arrangement.spacedBy((-124).dp),
+            userScrollEnabled = userScrollEnabled.value,
+        ) {
+            itemsIndexed(items) { index, item ->
+                MovieItem(
+                    item,
+                    index,
+                    listState,
+                    selectedCard,
+                    animationInProgress,
+                    userScrollEnabled,
+                    pagerState
+                )
+            }
+        }
+        AnimatedVisibility(visible = selectedCard.value != null) {
+            Column(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(top = 300.dp)
+            ) {
+                HorizontalPager(
+                    pageCount = 3,
+                    state = pagerState,
+                    modifier = Modifier
+                        .weight(1f)
+                        .fillMaxWidth()
+                ) {
+
+                        page ->
+                    Box(
+                        modifier = Modifier
+                            .fillMaxSize()
+                            .background(Color(0xFF, 0x80 * (page + 1), 0x80)),
+                        contentAlignment = Alignment.Center
+                    ) {
+                        Text(text = "Page $page", color = Color.White)
+                    }
+                }
+                PagerIndicator(
+                    pagerState = pagerState,
+                    pageCount = 3,
+                    modifier = Modifier.align(Alignment.CenterHorizontally)
+                )
+            }
+        }
+    }
+}
